Allow to configure item ratios in distributor filters
It allows to control the ratio in which certain items will use the different outputs.
This is done by adding an item multiple times to the same filter.

Because of this, the hardcoded permutations are replaced with a Fisher-Yates-Shuffle.
This is theoretically a little bit slower; but according to my measurements, about one million of copy-shuffle operations can be executed per second (tested with a table length of four); so it doesn't seem to be a big problem.

Additionally, a bit of outdated information was removed from the manuals.