Allow to pause the sequencer with a TechAge command
Has the same effect as turning the sequencer off by hand.

Use case: You use a sequencer in combination with a TA Signal Lamp to build a flashing warning light.
You then want to turn this warning light on and off with a TechAge command.
Using the "off" command would result in unsetting the "Run endless" checkbox.